Refactor loadEffectPreset function: Update effects ID

+++ examples/AlphaMatrixPixelsBase/effect_presets.hpp
@@ -23,7 +23,7 @@ using amp::csRenderMatrixCopy;
 // Abstract function: adds effects to the array based on effect ID
 // effectManager: reference to effect manager for adding effects (matrix is taken from effectManager.getMatrix())
 // effectId: ID of the effect to create
-void loadEffectPreset(csEffectManager& effectManager, uint8_t effectId, csMatrixPixels* matrixSecondBuffer = nullptr) {
+void loadEffectPreset(csEffectManager& effectManager, uint16_t effectId, csMatrixPixels* matrixSecondBuffer = nullptr) {
     if (effectId == 0) {
         return;
     }
